feat: test utilities for further test suite development
